requirements.txt
pyautogui
selenium
selenium-requests
opencv-python
numpy
python3_xlib
python-telegram-bot
cprofilev
psutil